Move CompetitionFile to competition.configs package
