+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/configs/CompetitionFile.java
@@ -1,8 +1,10 @@
-package com.oheers.fish.competition;
+package com.oheers.fish.competition.configs;
 
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.FishUtils;
 import com.oheers.fish.api.reward.Reward;
+import com.oheers.fish.competition.Bar;
+import com.oheers.fish.competition.CompetitionType;
 import com.oheers.fish.config.ConfigBase;
 import com.oheers.fish.config.messages.Message;
 import com.oheers.fish.fishing.items.FishManager;
